requirements.txt
google-generativeai>=0.3.0

python-dotenv>=1.0.0
pydantic>=2.0.0

pytest>=7.4.0
pytest-json-report>=1.5.0

pandas>=2.0.0
numpy>=1.24.0

matplotlib>=3.7.0
seaborn>=0.12.0
rich>=13.0.0

jsonschema>=4.17.0

markdown>=3.4.0
weasyprint>=59.0
